requirements-full.txt
-r requirements-medium.txt
opencv-python
sounddevice
numpy
openai-whisper
